Moved Quantity.hpp back to the proper position, updated code following change in Map and Matrix
