+++ src/resonanceReconstruction/rmatrix/CompoundSystem.hpp
@@ -14,8 +14,8 @@
 #include "range/v3/algorithm/for_each.hpp"
 #include "range/v3/view/all.hpp"
 #include "range/v3/view/transform.hpp"
-#include "resonanceReconstruction/quantities.hpp"
-#include "resonanceReconstruction/Map.hpp"
+#include "resonanceReconstruction/Quantity.hpp"
+#include "resonanceReconstruction/rmatrix/Map.hpp"
 #include "resonanceReconstruction/rmatrix/ReactionID.hpp"
 #include "resonanceReconstruction/rmatrix/ReactionChannelID.hpp"
 #include "resonanceReconstruction/rmatrix/Formalism.hpp"
