+++ src/resonanceReconstruction/rmatrix/LMatrixCalculator.hpp
@@ -7,8 +7,8 @@
 // other includes
 #include "range/v3/view/transform.hpp"
 #include "range/v3/view/zip_with.hpp"
-#include "resonanceReconstruction/quantities.hpp"
-#include "resonanceReconstruction/Matrix.hpp"
+#include "resonanceReconstruction/Quantity.hpp"
+#include "resonanceReconstruction/rmatrix/Matrix.hpp"
 #include "resonanceReconstruction/rmatrix/BoundaryOption.hpp"
 #include "resonanceReconstruction/rmatrix/Channel.hpp"
