+++ src/resonanceReconstruction/rmatrix/Particle.hpp
@@ -5,7 +5,7 @@
 
 // other includes
 #include "Log.hpp"
-#include "resonanceReconstruction/quantities.hpp"
+#include "resonanceReconstruction/Quantity.hpp"
 #include "resonanceReconstruction/rmatrix/ParticleID.hpp"
 #include "resonanceReconstruction/rmatrix/QuantumNumbers.hpp"
