+++ src/resonanceReconstruction/rmatrix/SpinGroup.hpp
@@ -24,8 +24,8 @@
 #include "range/v3/view/transform.hpp"
 #include "range/v3/view/zip_with.hpp"
 #include "range/v3/view/zip.hpp"
-#include "resonanceReconstruction/quantities.hpp"
-#include "resonanceReconstruction/Map.hpp"
+#include "resonanceReconstruction/Quantity.hpp"
+#include "resonanceReconstruction/rmatrix/Map.hpp"
 #include "resonanceReconstruction/rmatrix/ReactionChannelID.hpp"
 #include "resonanceReconstruction/rmatrix/Formalism.hpp"
 #include "resonanceReconstruction/rmatrix/BoundaryOption.hpp"
