+++ src/resonanceReconstruction/rmatrix/SpinGroup/test/SpinGroup.test.cpp
@@ -29,6 +29,7 @@ using ReactionChannelID = rmatrix::ReactionChannelID;
 using ShiftFactor = rmatrix::ShiftFactor;
 using Constant = rmatrix::Constant;
 using ReichMoore = rmatrix::ReichMoore;
+template < typename Key, typename Value > using Map = rmatrix::Map< Key, Value >;
 
 constexpr AtomicMass neutronMass = 1.008664 * daltons;
 constexpr ElectricalCharge elementary = dimwits::constant::elementaryCharge;
